Implement multitenancy structure Fixes #1 ✨feature: Implement multitenancy structure
