+++ apps/rallz_auth/managers.py
@@ -1,4 +1,6 @@
+from django.contrib.auth import models
 from django.contrib.auth.models import BaseUserManager
+from django.db import models
 
 
 class UserManager(BaseUserManager):
@@ -36,3 +38,18 @@ class UserManager(BaseUserManager):
         user.emailaddress_set.create(email=email, verified=True, primary=True)
         user.save()
         return user
+
+
+class OrgManager(models.Manager):
+    def get_for_user(self, user):
+        return self.get_queryset().filter(users=user)
+
+
+class ActiveOrgManager(OrgManager):
+    """
+    A more useful extension of the default manager which returns querysets
+    including only active organizations
+    """
+
+    def get_queryset(self):
+        return super().get_queryset().filter(is_active=True)
